backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
haystack-ai==2.0.1
python-multipart==0.0.17
python-dotenv==1.0.0
pydantic==2.5.0
python-docx==1.1.0
